+++ Coimbra/ILateUpdateService.cs
@@ -1,0 +1,26 @@
+﻿using UnityEngine.Scripting;
+
+namespace Coimbra
+{
+    /// <summary>
+    /// Provides easy access to Unity's LateUpdate callback without requiring the object to be an <see cref="UnityEngine.MonoBehaviour"/>.
+    /// </summary>
+    [RequireImplementors]
+    public interface ILateUpdateService
+    {
+        /// <summary>
+        /// Add a listener to the LateUpdate.
+        /// </summary>
+        void AddListener(ILateUpdateListener listener);
+
+        /// <summary>
+        /// Clear all listeners from the LateUpdate;
+        /// </summary>
+        void ClearListeners();
+
+        /// <summary>
+        /// Remove a listener from the LateUpdate.
+        /// </summary>
+        void RemoveListener(ILateUpdateListener listener);
+    }
+}